+++ packages/ui-app/src/SideMenu/types.ts
@@ -23,10 +23,12 @@ export interface DefaultMenuBarItem {
 	key: PageCategory['key']
 }
 
-export interface DefaultMenuProps {
+export interface BaseMenuProps {
 	barItems: DefaultMenuBarItem[]
 	getActivePageForCategory: (category: PageCategory['key']) => string
 	localCategory: PageCategory['key']
+}
+export interface DefaultMenuProps extends BaseMenuProps {
 	pageCategories: PageCategoryItems
 	pagesConfig: PagesConfigItems
 	renderMain: RenderSideMenuMain
